+++ plugins/watcher/package.js
@@ -1,6 +1,6 @@
 /* eslint-disable prefer-arrow-callback */
 Package.describe({
-    name: 'meteor-community:meteor-desktop-watcher',
+    name: 'communitypackages:meteor-desktop-watcher',
     version: '2.2.6',
     summary: 'Watches .desktop dir and triggers rebuilds on file change.',
     git: 'https://github.com/Meteor-Community-Packages/meteor-desktop',
@@ -16,7 +16,7 @@ Package.onUse(function onUse(api) {
     api.versionsFrom('METEOR@1.3.3');
     api.use('ecmascript');
     api.use([
-        'meteor-community:meteor-desktop-bundler@2.2.6',
+        'communitypackages:meteor-desktop-bundler@2.2.6',
     ], ['server'], {
         weak: true
     });